fix: update message type from Float32 to Float64 for joint control publisher

+++ src/nhk2026_control/include/nhk2026_control/ide_arm_sim_bridge.hpp
@@ -3,7 +3,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp/publisher.hpp"
 
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/float64.hpp"
 #include "std_msgs/msg/float32_multi_array.hpp"
 
 class IdeArmSimBridge
@@ -12,7 +12,7 @@ class IdeArmSimBridge
 public:
     IdeArmSimBridge();
 private:
-    std::vector<rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr> joint_control_publisher_;
+    std::vector<rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr> joint_control_publisher_;
     
     void joint_array_control_callback(const std_msgs::msg::Float32MultiArray::SharedPtr rxdata, int index);
     std::vector<rclcpp::Subscription<std_msgs::msg::Float32MultiArray>::SharedPtr> control_subscriber_;
